+++ app/assets/stylesheets/findingaids.css.scss
@@ -2,7 +2,7 @@
 $panel-default-heading-bg: #fff;
 $grid-float-breakpoint: 980px;
 
-@import "nyulibraries";
+@import "nyulibraries_stylesheets/nyulibraries";
 @import "blacklight";
 
 @import "contenttypes/*.png";
